requirements.txt
PyQt5==5.15.10
pyqtgraph==0.13.3
numpy==1.24.3
pyserial==3.5
pymodbus==3.5.4
websockets==12.0
aiohttp==3.9.1
requests==2.31.0
pytest==7.4.3
pytest-qt==4.2.0
win10toast; sys_platform == "win32"
# Linux: notify-send is a system command (not a Python package)
#        Install with: sudo apt-get install libnotify-bin (Ubuntu/Debian)
#        Or: sudo yum install libnotify (RedHat/CentOS)
